% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{contar_dados}
\alias{contar_dados}
\title{La función contar_dados}
\usage{
contar_dados(tirada, x)
}
\arguments{
\item{tirada}{vector numérico de largo \code{n} con la salida de la función \code{tirar_dados(n)}.}

\item{x}{valor entero entre 1 y 6 a ser contado.}
}
\value{
vector numérico de largo 1 con la cantidad de veces que salió el número \code{x} en \code{tirada}.
}
\description{
Esta función permite contar cuántos dados de valor \code{x} salieron en una \code{tirada}.
}
\examples{
nueva_tirada = tirar_dados(5)
mostrar_dados(nueva_tirada)
contar_dados(nueva_tirada, 1)

}
