% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{leer_opciones}
\alias{leer_opciones}
\title{La función leer_opciones}
\usage{
leer_opciones(titulo, ...)
}
\arguments{
\item{titulo}{cadena de texto que se imprime al inicio.}

\item{...}{cadenas de texto con cada una de las opciones.}
}
\value{
vector numérico de largo 1 con la posición de la opción elegida.
}
\description{
Esta función permite mostrar varias opciones y leer una de ellas.
}
\details{
el valor de salida respeta el orden en el pasaje de argumentos.
}
\examples{
leer_opciones("Quien es el mas grande?", "Central", "Newell's")

}
