% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{leer_palabra}
\alias{leer_palabra}
\title{La función leer_palabra}
\usage{
leer_palabra()
}
\value{
vector caracter de largo 1 con la palabra leída.
}
\description{
La función leer_palabra
}
\examples{
leer_palabra()

}
