% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{limpiar_consola}
\alias{limpiar_consola}
\title{La función limpiar_consola}
\usage{
limpiar_consola()
}
\value{
el objeto \code{NULL}.
}
\description{
Una función para limpiar la consola.
}
\examples{
limpiar_consola()

}
