% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{mostrar_dados}
\alias{mostrar_dados}
\title{La función mostrar_dados}
\usage{
mostrar_dados(tirada)
}
\arguments{
\item{tirada}{vector numérico de largo \code{n} con la salida de la función \code{tirar_dados(n)}.}
}
\value{
vector caracter de largo 1 con el resultado de la tirada.
}
\description{
Esta función muestra los dados resultantes de una tirada.
}
\examples{
nueva_tirada = tirar_dados(5)
mostrar_dados(nueva_tirada)

}
