% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{pausa}
\alias{pausa}
\title{La función pausa}
\usage{
pausa(mensaje = "Pausa del juego.")
}
\arguments{
\item{mensaje}{cadena de texto con el mensaje a mostrar.}
}
\value{
el objeto \code{NULL}.
}
\description{
Esta función pausa la ejecución de código hasta que el usuario ingrese la letra (c).
}
\examples{
pausa()

}
