% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{texto_lento}
\alias{texto_lento}
\title{La función texto_lento}
\usage{
texto_lento(..., pausa = 0.02)
}
\arguments{
\item{...}{cadenas de texto a imprimir.}

\item{pausa}{valor numérico con el tiempo de impresión por caracter en segundos.}
}
\value{
el objeto \code{NULL}.
}
\description{
Esta función imprime texto lentamente.
}
\details{
Las cadenas de texto se concatenan con espacios, similar al uso de \code{cat()}.
}
\examples{
texto_lento(5, "x", 7, "es igual a", 5*7)
texto_lento("Un mensaje demasiado lento", pausa = 0.2)

}
