% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{tirar_dados}
\alias{tirar_dados}
\title{La función tirar_dados}
\usage{
tirar_dados(n)
}
\arguments{
\item{n}{entero positivo con el número de dados a ser tirados.}
}
\value{
vector numérico de largo \code{n} con el resultado de la tirada.
}
\description{
Esta función permite simular la tirada de un número arbitrario de dados legales de seis caras.
}
\examples{
tirar_dados(5)

}
