% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{titulo}
\alias{titulo}
\title{La función titulo}
\usage{
titulo(texto, ancho = 70)
}
\arguments{
\item{texto}{cadena de texto con el título a mostrar.}

\item{ancho}{valor numérico con el ancho de las ligas que encierran al título.}
}
\value{
el objeto \code{NULL}.
}
\description{
Esta función imprime una cadena de texto en formato de encabezado.
}
\examples{
titulo("NUEVA RONDA")

}
